requirements.txt
boto3==1.5.6
botocore==1.8.20
Django==2.0
djangorestframework==3.7.7
docutils==0.14
jmespath==0.9.3
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
httpie==1.0.0
